At several places LONGEST and int/long were mixed. This is fine on a 64 bits machine or a 32 bits machine without "long long" implemented. Since gcc now supports long long on many platforms...
With constructors and the like it is difficult to detect the first stack
frame. GDB often goes too far and the associated command crashes before
it can do its stuff. Thus, tm-i386.h and corefile were modified to at
least stop when encountering a "silly" frame pointer.

A Register map was defined in m3-lang.c to activate the thread switch command
on Linux.

Some symbols in shared libraries (ELF) were not processed correctly
and GDB would have the wrong address for global variables. Because of
this the thread and type information accessed through RT0u as a global
variable would be unaccessible.

In symfile.c the fixups of modula-3 interface names were not always
applied correctly.