From SRC patches: bug in updating heap statistics (09-Feb-98)
This bug applies to all platforms.

There is a bug in the code that computes statistics about runtime heap
usage. We suspect this bug was causing shownew(1) to report incorrect results.

The fix is simple. Edit the file "m3/m3core/src/runtime/common/RTutils.m3". In
the Delta procedure around line 76, change the WITH statement from:

      WITH a = v.r[i], b = v2.r[i] DO
        DEC (a.total.count, b.total.count);
        DEC (a.total.size,  b.total.size);
      END;

to

      WITH x = v.r[i].total, a = v1.r[i].total, b = v2.r[i].total DO
        x.count := a.count - b.count;
        x.size  := a.size  - b.size;
      END;

Then rebuild and ship the m3core package.

Thanks to John Kominek for pointing out the bug and providing the fix. This
bug has been fixed in all releases of SRC Modula-3 after 3.6.